+++ src/tests/stress/test_latency_measurement.cpp
@@ -454,7 +454,8 @@ TEST_F(LatencyMeasurementTest, AudioLatency) {
         uint64_t send_time_us = last_high_energy_send_time_us.load();
 
         if (send_time_us > 0) {
-          double latency_ms = (receive_time_us - send_time_us) / 1000.0;
+          double latency_ms =
+              static_cast<double>(receive_time_us - send_time_us) / 1000.0;
           if (latency_ms > 0 && latency_ms < 5000) { // Sanity check
             stats.addMeasurement(latency_ms);
             std::cout << "  Audio latency: " << std::fixed
@@ -676,7 +677,8 @@ TEST_F(LatencyMeasurementTest, FullDeplexAudioLatency) {
       if (send_from_a_us == 0 || detect_us <= send_from_a_us) {
         continue;
       }
-      double a_to_b_ms = (detect_us - send_from_a_us) / 1000.0;
+      double a_to_b_ms =
+          static_cast<double>(detect_us - send_from_a_us) / 1000.0;
       if (a_to_b_ms < kMinValidOneWayMs || a_to_b_ms > 5000) {
         continue;
       }
@@ -742,7 +744,8 @@ TEST_F(LatencyMeasurementTest, FullDeplexAudioLatency) {
       }
       a_received_pulse_id.store(pulse_id);
 
-      double b_to_a_ms = (receive_us - send_from_b_us) / 1000.0;
+      double b_to_a_ms =
+          static_cast<double>(receive_us - send_from_b_us) / 1000.0;
       if (b_to_a_ms >= kMinValidBToAMs && b_to_a_ms < 5000) {
         b_to_a_stats.addMeasurement(b_to_a_ms);
         std::cout << "  B->A latency: " << std::fixed << std::setprecision(2)
@@ -750,7 +753,8 @@ TEST_F(LatencyMeasurementTest, FullDeplexAudioLatency) {
       }
 
       if (send_from_a_us > 0) {
-        double rtt_ms = (receive_us - send_from_a_us) / 1000.0;
+        double rtt_ms =
+            static_cast<double>(receive_us - send_from_a_us) / 1000.0;
         if (rtt_ms > 0 && rtt_ms < 10000) {
           round_trip_stats.addMeasurement(rtt_ms);
           std::cout << "  A->B->A latency: " << std::fixed
